Create Generalized Least Squares Regression
This code fits a known nanodisc spectra and a known IDP spectra to an experimental spectra to determine the relative amounts of sencondary structures and better understand aSyn's conformation in cell. The data is read in as a CSV file containing chemical shift, nanodisc data, IDP data, and experimental (target) data. The nanodisc and IDP data are set to each be an input variable for a linear regression while the experimental data is the "target" we are tryng to simulate. A generalized least squares algorithm from statsmodels.api is then run. The output of this regression is a table of coefficients (weights)and standard error, along with an R^2 value. The data is then weighted by its corresponding coefficient and the area under each curve is calculated through a trapezoidal approximation. The percentage of each structure represented is the contribution that the integral of the respective data makes toward the best fit prediction. Lastly, a graph is made with the nanodisc data, IDP data, experimental data, and best fit approximation.